requirements.txt
asgiref==3.11.1
Django==5.2.13
djangorestframework==3.17.1
psycopg[binary]==3.2.9
sqlparse==0.5.5
